FIX-1985 remove VirtualizedContentTable
