+++ Src/WitsmlExplorer.Frontend/components/ContentViews/table/index.ts
@@ -1,3 +1,2 @@
-export * from "./tableParts";
 export * from "./ContentTable";
-export * from "./VirtualizedContentTable";
+export * from "./tableParts";
